refactor actual container stuff into container.c (closes: #170, closes: #187)
